+++ server/routes/auth.js
@@ -1,0 +1,11 @@
+var express = require('express');
+var router = express.Router();
+var authRouter = require("../controllers/auth.js");
+
+router.get('/acceso', authRouter.logInIndex);
+router.post('/acceso', authRouter.logIn);
+
+router.get('/registro', authRouter.signUpIndex);
+router.post('/registro', authRouter.signUp);
+
+module.exports = router;
