+++ views/index.jade
@@ -4,27 +4,27 @@ extends layout
 block append head
   link(rel="stylesheet" href="stylesheets/ammap.css" type="text/css")
   script(src="js/ammap.js" type="text/javascript")
-  script(src="js/mexicoHigh.js" type="text/javascript") 
-  
+  script(src="js/mexicoHigh.js" type="text/javascript")
+
 
 
 block content
   .container
     nav.navbar.navbar-expand-lg.navbar-light.bg-light
-      a.navbar-brand(href='#') 
+      a.navbar-brand(href='#')
         img.img-fluid(src='/images/icon.png', alt='Red de Investigadores')
       button.navbar-toggler(type='button' data-toggle='collapse' data-target='#navbarNav' aria-controls='navbarNav' aria-expanded='false' aria-label='Toggle navigation')
         span.navbar-toggler-icon
       .collapse.navbar-collapse(id='navbarNav')
         ul.navbar-nav.mr-auto
           li.nav-item.active
-            a.nav-link(href='#') 
+            a.nav-link(href='#')
               button.btn.btn-primary(type='button')
                 strong Inicio
-                span.sr-only (current)    
-          if !authenticated 
+                span.sr-only (current)
+          if !authenticated
             li.nav-item
-              a.nav-link(href='/logIn')
+              a.nav-link(href='/acceso')
                 button.btn.btn-primary
                   strong Iniciar Sesión
           if !authenticated
@@ -35,14 +35,10 @@ block content
           if authenticated
             li.nav-item
               a.nav-link(href='/perfil')
-                button.btn.btn-primary 
+                button.btn.btn-primary
                   strong Perfil
 
     .map
-      
+
       script(src="js/drawMap.js")
       div(id="map" style="width: 100%; background-color:#ffffff; height: 500px;")
-      
-      
-    
-
